Java Programming
Create a class Account with the variables to represent the account holder’s name, account number, account type (S for savings and C for Current), balance amount, use constructor to initialize the variables. Write method to deposit money, withdrawal of money, mutator and accessor methods, toString method to return object as a string. Use static variable to keep track the number of transactions. Create Driver class to demonstrate the functioning of the above.